+++ Src/Server/MK.Booking/MK.Booking/ReadModel/PaymentInformationDetails.cs
@@ -10,8 +10,8 @@ namespace apcurium.MK.Booking.ReadModel
 
         public Guid? CreditCardId { get; set; }
 
-        public decimal? TipAmount { get; set; }
+        public double? TipAmount { get; set; }
 
-        public decimal? TipPercent { get; set; }
+        public double? TipPercent { get; set; }
     }
 }
